@@ -4,28 +4,31 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=3.10.0 <4.0.0"
 
 dependencies:
-  cached_network_image: ^3.2.1
-  firebase_messaging: ^11.3.0
+  cached_network_image: ^3.4.1
+  dynamic_color: ^1.8.1
+  firebase_core: ^4.3.0
+  firebase_messaging: ^16.1.0
   flutter:
     sdk: flutter
-  flutter_html: ^2.2.1
-  flutter_share: ^2.0.0
-  hive: ^2.1.0
+  flutter_html: ^3.0.0
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  http: ^0.13.4
+  http: ^1.6.0
   intl: ^0.17.0
-  provider: ^6.0.2
-  url_launcher: ^6.1.2
+  provider: ^6.1.5
+  share_plus: ^12.0.1
+  url_launcher: ^6.3.2
   webfeed: ^0.7.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^6.0.0
   flutter_launcher_icons: "^0.9.2"
+  import_sorter: ^4.6.0
 
 flutter_icons:
   android: true
