@@ -1,17 +1,19 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter_html/flutter_html.dart';
-import 'package:flutter_share/flutter_share.dart';
+import 'package:share_plus/share_plus.dart';
+import 'package:url_launcher/url_launcher_string.dart';
+
 import 'package:islieb/configs/app_assets.dart';
 import 'package:islieb/configs/app_constants.dart';
 import 'package:islieb/configs/app_themes.dart';
 import 'package:islieb/model/islieb_reader.dart';
 import 'package:islieb/utils/localized_date.dart';
-import 'package:url_launcher/url_launcher_string.dart';
 
 class HomeView extends StatefulWidget {
-  const HomeView({Key? key}) : super(key: key);
+  const HomeView({super.key});
 
   @override
   State<HomeView> createState() => _HomeViewState();
@@ -28,13 +30,12 @@ class _HomeViewState extends State<HomeView> {
           (IsliebReader.of(context).rssFeed?.items?.length ?? 0);
 
   void _updateDisplayBackButton() => setState(() {
-        displayBackButton =
-            _pageController.hasClients && _pageController.page != 0;
-      });
+    displayBackButton = _pageController.hasClients && _pageController.page != 0;
+  });
 
   void _updateAction() => setState(() {
-        IsliebReader.of(context).loadRssFeed();
-      });
+    IsliebReader.of(context).loadRssFeed();
+  });
 
   @override
   void initState() {
@@ -55,29 +56,21 @@ class _HomeViewState extends State<HomeView> {
   }
 
   void _infoButtonAction() => showAboutDialog(
-        context: context,
-        applicationName: AppConstants.applicationName,
-        applicationIcon: Image.asset(
-          AppAssets.icon,
-          width: 64,
-          height: 64,
-        ),
-        children: [
-          Html(
-            style: AppThemes.htmlStyle,
-            onLinkTap: (String? url, RenderContext context,
-                    Map<String, String> attributes, _) =>
-                url == null
-                    ? null
-                    : launchUrlString(
-                        url,
-                        mode: LaunchMode.externalApplication,
-                      ),
-            data:
-                '<p>Comics and contents by <a href="https://islieb.de">islieb</a>.</p><p>App created with love by <a href="https://krillefear.gitlab.io">Krille Fear</a>.</p>',
-          ),
-        ],
-      );
+    context: context,
+    applicationName: AppConstants.applicationName,
+    applicationIcon: Image.asset(AppAssets.icon, width: 64, height: 64),
+    children: [
+      Html(
+        style: AppThemes.htmlStyle,
+        onLinkTap: (String? url, Map<String, String> attributes, _) =>
+            url == null
+            ? null
+            : launchUrlString(url, mode: LaunchMode.externalApplication),
+        data:
+            '<p>Comics and contents by <a href="https://islieb.de">islieb</a>.</p><p>App created with love by <a href="https://krille-chan.github.io/">Krille</a>.</p>',
+      ),
+    ],
+  );
 
   @override
   Widget build(BuildContext context) {
@@ -89,9 +82,7 @@ class _HomeViewState extends State<HomeView> {
         return items == null
             ? Scaffold(
                 appBar: AppBar(title: const Text('Lade ...')),
-                body: const Center(
-                  child: CircularProgressIndicator.adaptive(),
-                ),
+                body: const Center(child: CircularProgressIndicator.adaptive()),
               )
             : Scaffold(
                 body: PageView(
@@ -109,14 +100,19 @@ class _HomeViewState extends State<HomeView> {
                               ),
                             IconButton(
                               icon: Icon(Icons.adaptive.share_outlined),
-                              onPressed: () => FlutterShare.share(
-                                title:
-                                    item.title ?? AppConstants.applicationName,
-                                linkUrl:
+                              onPressed: () => SharePlus.instance.share(
+                                ShareParams(
+                                  title:
+                                      item.title ??
+                                      AppConstants.applicationName,
+                                  uri: Uri.tryParse(
                                     (item.content?.images.isNotEmpty ?? false
                                             ? item.content?.images.first
                                             : null) ??
-                                        item.link,
+                                        item.link ??
+                                        '',
+                                  ),
+                                ),
                               ),
                             ),
                             IconButton(
@@ -125,65 +121,60 @@ class _HomeViewState extends State<HomeView> {
                             ),
                           ],
                         ),
-                        body: SingleChildScrollView(
-                          child: Html(
-                            data: item.content?.value ?? '',
-                            onLinkTap: (String? url, RenderContext context,
-                                    Map<String, String> attributes, _) =>
-                                url == null
-                                    ? null
-                                    : launchUrlString(
-                                        url,
-                                        mode: LaunchMode.externalApplication,
-                                      ),
-                            style: AppThemes.htmlStyle,
-                            customRender: {
-                              'img': (context, parsedChild) =>
-                                  CachedNetworkImage(
-                                    imageUrl: context.tree.attributes['src']!,
+                        body: RefreshIndicator.adaptive(
+                          onRefresh: () => isliebReader.loadRssFeed(),
+                          child: SingleChildScrollView(
+                            child: Html(
+                              data: item.content?.value ?? '',
+                              onLinkTap:
+                                  (
+                                    String? url,
+                                    Map<String, String> attributes,
+                                    _,
+                                  ) => url == null
+                                  ? null
+                                  : launchUrlString(
+                                      url,
+                                      mode: LaunchMode.externalApplication,
+                                    ),
+                              style: AppThemes.htmlStyle,
+                              extensions: [
+                                TagExtension(
+                                  tagsToExtend: {'img'},
+                                  builder: (context) => CachedNetworkImage(
+                                    imageUrl: context.attributes['src']!,
                                     fit: BoxFit.fitWidth,
-                                    errorWidget: (_, __, ___) => const SizedBox(
+                                    errorWidget: (_, _, _) => const SizedBox(
                                       height: 256,
                                       child: Center(
                                         child: Icon(Icons.wifi_off_outlined),
                                       ),
                                     ),
-                                    placeholder: (_, __) => const SizedBox(
+                                    placeholder: (_, _) => const SizedBox(
                                       height: 256,
                                       child: Center(
-                                        child: CircularProgressIndicator
-                                            .adaptive(),
+                                        child:
+                                            CircularProgressIndicator.adaptive(),
                                       ),
                                     ),
                                   ),
-                              'audio': (context, parsedChild) => OutlinedButton(
+                                ),
+                                TagExtension(
+                                  tagsToExtend: {'audio', 'video', 'iframe'},
+                                  builder: (context) => OutlinedButton(
                                     onPressed: () => launchUrlString(
-                                      context.tree.attributes['src']!,
+                                      context.attributes['src']!,
                                       mode: LaunchMode.externalApplication,
                                     ),
                                     child: const Text('Audio abspielen'),
                                   ),
-                              'video': (context, parsedChild) => OutlinedButton(
-                                    onPressed: () => launchUrlString(
-                                      context.tree.attributes['src']!,
-                                      mode: LaunchMode.externalApplication,
-                                    ),
-                                    child: const Text('Video abspielen'),
-                                  ),
-                              'iframe': (context, parsedChild) =>
-                                  OutlinedButton(
-                                    onPressed: () => launchUrlString(
-                                      context.tree.attributes['src']!,
-                                      mode: LaunchMode.externalApplication,
-                                    ),
-                                    child:
-                                        const Text('Externen Inhalt anzeigen'),
-                                  ),
-                            },
+                                ),
+                              ],
+                            ),
                           ),
                         ),
                       );
-                    }).toList(),
+                    }),
                     Scaffold(
                       appBar: AppBar(
                         title: const Text(AppConstants.applicationName),
@@ -195,111 +186,70 @@ class _HomeViewState extends State<HomeView> {
                         ],
                       ),
                       body: Column(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            Material(
-                              elevation: 2,
-                              borderRadius: BorderRadius.circular(12),
-                              clipBehavior: Clip.hardEdge,
-                              child: Image.asset(
-                                AppAssets.icon,
-                                width: 128,
-                                height: 128,
-                              ),
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        children: [
+                          Material(
+                            elevation: 2,
+                            borderRadius: BorderRadius.circular(12),
+                            clipBehavior: Clip.hardEdge,
+                            child: Image.asset(
+                              AppAssets.icon,
+                              width: 128,
+                              height: 128,
                             ),
-                            const Padding(
-                              padding: EdgeInsets.all(12.0),
-                              child: Text(
-                                'Huch?! Es wurden keine weiteren Comics gefunden. Schau vielleicht mal auf der Website!',
-                                textAlign: TextAlign.center,
-                                style: TextStyle(
-                                  fontSize: 18,
-                                ),
-                              ),
+                          ),
+                          const Padding(
+                            padding: EdgeInsets.all(12.0),
+                            child: Text(
+                              'Huch?! Es wurden keine weiteren Comics gefunden. Schau vielleicht mal auf der Website!',
+                              textAlign: TextAlign.center,
+                              style: TextStyle(fontSize: 18),
                             ),
-                            OutlinedButton.icon(
-                              onPressed: () => launchUrlString(
-                                AppConstants.website,
-                                mode: LaunchMode.externalApplication,
-                              ),
-                              label: const Text(AppConstants.website),
-                              icon: const Icon(Icons.open_in_new),
+                          ),
+                          OutlinedButton.icon(
+                            onPressed: () => launchUrlString(
+                              AppConstants.website,
+                              mode: LaunchMode.externalApplication,
                             ),
-                          ]),
+                            label: const Text(AppConstants.website),
+                            icon: const Icon(Icons.open_in_new),
+                          ),
+                        ],
+                      ),
                     ),
                   ],
                 ),
-                bottomNavigationBar: Material(
-                  elevation: 20,
-                  color: Theme.of(context).appBarTheme.backgroundColor,
+                floatingActionButtonLocation:
+                    FloatingActionButtonLocation.miniCenterDocked,
+                floatingActionButton: Padding(
+                  padding: const EdgeInsets.symmetric(horizontal: 8.0),
                   child: Row(
                     children: [
-                      if (displayBackButton) ...[
-                        IconButton(
-                          padding: EdgeInsets.zero,
-                          icon: const Icon(Icons.first_page_outlined),
-                          color: Theme.of(context).colorScheme.primary,
-                          splashRadius: 16,
-                          onPressed: () => _pageController.animateToPage(
-                            0,
-                            duration: const Duration(milliseconds: 200),
-                            curve: Curves.bounceInOut,
-                          ),
-                        ),
-                        TextButton(
-                          onPressed: () => _pageController.previousPage(
-                            duration: const Duration(milliseconds: 200),
-                            curve: Curves.bounceInOut,
-                          ),
-                          child: Row(
-                            children: const [
-                              Icon(Icons.chevron_left),
-                              Text('Neuer'),
-                            ],
-                          ),
-                        ),
-                      ],
-                      if (!displayBackButton)
-                        TextButton(
-                          onPressed:
-                              snapshot.connectionState != ConnectionState.done
-                                  ? null
-                                  : _updateAction,
-                          child: Row(
-                            children: [
-                              snapshot.connectionState != ConnectionState.done
-                                  ? const Padding(
-                                      padding: EdgeInsets.all(4.0),
-                                      child: SizedBox(
-                                          width: 16,
-                                          height: 16,
-                                          child: CircularProgressIndicator
-                                              .adaptive(
-                                            strokeWidth: 2,
-                                          )),
-                                    )
-                                  : const Icon(Icons.refresh_outlined),
-                              const Text('Aktualisieren'),
-                            ],
-                          ),
-                        ),
-                      const Spacer(),
-                      TextButton(
-                        onPressed: (!_pageController.hasClients &&
-                                    items.isNotEmpty) ||
-                                canSwipe
-                            ? () => _pageController.nextPage(
+                      if (displayBackButton)
+                        FloatingActionButton(
+                          mini: true,
+                          onPressed: displayBackButton
+                              ? () => _pageController.animateToPage(
+                                  0,
                                   duration: const Duration(milliseconds: 200),
                                   curve: Curves.bounceInOut,
                                 )
-                            : null,
-                        child: Row(
-                          children: const [
-                            Text('Älter'),
-                            Icon(Icons.chevron_right),
-                          ],
+                              : null,
+                          child: const Icon(Icons.chevron_left),
                         ),
+                      Spacer(),
+                      FloatingActionButton(
+                        mini: true,
+                        onPressed:
+                            (!_pageController.hasClients && items.isNotEmpty) ||
+                                canSwipe
+                            ? () => _pageController.nextPage(
+                                duration: const Duration(milliseconds: 200),
+                                curve: Curves.bounceInOut,
+                              )
+                            : null,
+                        child: Icon(Icons.chevron_right),
                       ),
                     ],
                   ),
