@@ -1,6 +1,5 @@
-package de.islieb.islieb
+package de.islieb.app
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()
